Add logging role
The logging role configures Logstash 1.5 with Docker and systemd.
Containers with Logstash are started on each host and collect logs from
multiple inputs. This allows us to filter all logs and send them to a
central location. At this time, we support several outputs such as HDFS
and Logentries. By default all logs are sent to STDOUT.