zk container - manage zk cls with zkcli and add certs
* uses new version of zookeeper container with zookeepercli built in
* uses zookeepercli to set zk acls instead of the setacl.sh
  script (https://github.com/CiscoCloud/microservices-infrastructure/issues/329).
* zookeeper-service systemd unit has been updated to use the
  certificates in /etc/consul/ssl for ssl validation.

Note that the zookeeper container includes an updated version of
consul-template (v0.10.0) which has a
bug (https://github.com/hashicorp/consul-template/issues/321) where it
ignores some command line flags such as `--ssl-verify`. To mitigate
this (and for better ssl support in general), the zookeeper container
now allows you to mount a host volume with certificates to use for ssl
verification.