All consul checks must have timeouts
Previously, the checks were defined without
a timeout, which led to thousands of processes
spawned by consul. This would worsen the
health of an already unstable cluster.

In particular, the curl command by default has
no timeout, and will hang if it successfully
opens a TCP connection but receives no data.