Support for DNS providers
Every cloud provider exposes IPs for the
worker and control nodes. These IPs are
hooked up to the DNS providers, to map
IPs to DNS records.

This way, to cloud providers and the DNS
providers are loosely coupled.

There is a reference DNS module for dnsimple.
Also added is an example of how to hook it
up to the existing cloud providers.