Rewrite docker config files to enable deployments
Right now, users have to install terraform and ansible, and their
dependencies, in order to user Mantl. Vagrant is a demo/test-drive
option, but not for other use cases. After the installation(s), then
they have to run multiple commands in order to get a cluster up.

What I'm working towards is a way for users to be able to get started
with Mantl in one or two steps, that removes the need to install
dependenies. This commit reworks Dockerfile and docker_launch.sh, and in
the tests that I have done, it can accomplish this.

Using an alpine linux base for this image reduces the size of the image,
although we do have to add build utils back in for the python
dependencies. As we reduce those dependencies, this image will get
smaller.

I added a MANTL_CONFIG_DIR env var to the dockerfile, and set it to
/local. This directory is there to hold custom configuration for users.
It has to be different than /mantl (where the source code goes) because
we don't want users to have to clone the repository in order to add in
their custom files. We add the `/bin/sh -c` entrypoint to make this
image closer to an executable.

Removing the ONBUILD statements from the Dockerfile simplifies using
this image as a base to build other images from. They are replaced with
logic in docker_launch.sh to check for user-provided files in a
particular place.  If these files don't exist, then we copy/generate the
files. When users mount a host volume in the MANTL_CONFIG_DIR, they will
get copies of all of the generated files in that volume, which they are
free to customize further. Documentation will need to be added for what
needs to be mounted where.

TERRAFORM_STATE_ROOT is changed to TERRAFORM_STATE. Because the
expectation is that users will mount the config dir onto their local
machines, they will want the tfstate file persist between docker runs.