Add logging, set global docker envs in travis
When I ported the setup logic to python, I did not adding any logging
output. I have fixed that, by including helpful messages to users on
what the setup is currently doing. I also have the setup script set the
SSH_KEY env if none is set, instead of aborting.

I was forgetting to add settings to multiple places, so I figured that
it was time for some DRY refactoring. This adds global envs for the
docker run lines, to make it easier to keep things consistent.

For example, I changed docker run to attach stdout & stderr instead of
allocating a pseudo-tty. This change allows the security setup script to
generate credentials randomly, as they did before.

The volume parts of the docker run call need to be placed on the same
lines, because that's the most direct way to ensure that they are
populated to the correct values. If there is a DRY way of doing this,
that would be better, but I can't think of anything right now.

During my heavy edits of the travis file, I found that the indentation
was inconsistent. I ensured that my editor had the correct tab/space
settings, and adjusted the indentation.