pull request template: remove rebase requirement
This should still be a requirement, but it's not helpful to have in the
template because it can't really be evaluated until after all discussion
and testing is finished. Having another box to check also means that in
Github's UI there are a lot of issues that have 3/4 requirements
checked where the fourth is the rebase requirement, and which are
otherwise completely ready for merge.