Fix symlink/env issues, enable ssh-agent & destroy
I was on the wrong path when I was working on this last
night. I found a better way of ensuring that the symlink target is
correctly assigned: using the os.path.join function instead of the
abspath/relpath functions. This fixes the symlink logic for terraform_files.

After that, I was researching the docker run `--env` flag, I discovered this
line in the official docker docs:

> [I]f no `=` is provided, then that variable's current value is passed
> through

I just have to pass in the names of the secret vars, and
docker should pull them in. Until I found this particular solution, I
spent many commits trying to debug this problem.

Because ssh-agent and other ssh-related code has to be run often in the
docker containers, I moved the needed invocations to a separate script,
docker_ssh.sh. This will make docker run invocations shorter.

Because I prepended this to the build-cluster command, I removed
ssh keygen logic from build-cluster.
We need ssh keys to be symlinked for launch and testing, but we don't
want to run the whole setup script each time.
Assuming that the keys have already been generated from a run of
docker_setup.py, all that function does is symlink them.

The destroy steps need to be executed in the docker container as well,
even though I have not been able to get to this point in my testing.