Run ssh-agent as entrypoint, fix relpath in setup
ssh-agent has a 'lifetime option' where you can pass in a command and
args. The agent then only runs for a specified time, and the command is
run in a subshell. With the /bin/sh -c invocation at the start of the
ssh-agent's subshell, the image remains flexible.

The other change in here is that I fixed the relpath for the
docker_setup script in destroy_build.sh. With this change, the builds
will be destroyed properly once again.