Rename docker_setup -> docker, add cli-type args
We are rolling setup and deploy commands into one python script,
docker.py. It doesn't have much functionality, but this PR is getting
out of scope quickly, and it's needed for CI.

The destroy_build and start_build scripts were just wrappers around a
couple of shell commands, which can be wrapped by docker.py instead. I
also had logic in the security_setup wrapper function to move the
generated security file into MANTL_CONFIG_DIR, then symlink it. However,
Travis threw an error. Since that functionality is for users, and not
CI, I elected to remove that functionality, and add it in later if
needed.

To optimize for the size of the docker image, I have apk delete the
package. curl is a dependency of git, so when I removed that package, I
also removed curl. Therefore, we install it separately.

I checked out run-tests.sh from testing/switchboard, and ported the
documentation skipping, and skipping on master pushes.