Add travis user for OpenStack testing
The travis user was created for the shipped cluster, so now we can use
that username and a new key. Now we can run actual testing code on OS.

With the key I was using previously, I was running into a problem where
ssh asked for a passphrase, even though I created the key without one.
It seems the problem was from when the file was converted to a env
string then back to a file. I'm using travis's file encryption feature
to add the ssh key directly. Additionally, I disabled strict host key
checking, added the remote ssh command logic to ci_destroy.