TRAVIS: Add pub key to testing/gce.tf, readd edits
For a while now, when GCE is tested in travis, the remote-exec step
would fail. After much tinkering, I found that all I needed to do was to
edit the testing/gce.tf file to point to the pub ssh key generated
during CI runs, and it worked on my local machine.

There were a few edits here that seems to have been lost during the
rebase. A formatting change and a line edit isn't bad to loose in
squashing about 200 commits!