Update testing/gce.tf file to use modules
I was comparing the gce terraform file that I made for testing to the
most recent gce.sample.tf file. It turns out that I never updated the
file to use the modules! I need to work on getting the testing system to
test against the sample files instead of custom ones.