Separate setup/build for OpenStack, logfmt change
The ci-build logic for OS included a lot of setup. I took out the setup
logic and put it in ci-setup, and I added a call to ln to create the
symlink for the terraform file that I want to use for the test. (I
changed the matrix in .travis.yml to specify the actual terraform file
to use.) Also, I've added logging statements to track the final ssh
command sent to the jump box.

Also, I changed the format for logs. The time should come first