Improve OS ssh commands, add empty TESTING file
During my local testing for OpenStack (running the docker image with
travis hardcoded), I found that the TF var `build_number` was not being
defined properly. I spent quite a lot of time debugging this, but I
believe that I have fixed this, by creating a tfvar file with the
information in it. If travis fails again, because other vars aren't
defined in the environment, I'll have a better handle of how to fix it.

Additionally, I have created a TESTING file. The reason why I did this
was because, for some reason, when a `cd` command was the first line of
the ssh command, the directory would not actually be changed. I'm not
sure why that is, but I added in commands to create a TESTING file to
help work around that, but also to signal to potentially other people
when a build is in progess on the server.