Remove references to OS_KEY, check OS_IP instead
When I moved the ssh key from environment variable to encrypted file in
travis, I did not remove references to it in `.travis.yml` nor from
`docker.py`. This was causing the check for OS to fail, and not
triggering the remote commands. I have replaced these checks with OS_IP,
which is still in the environment.