Get head commit, add/kill OS ssh-agent, scrubbing
Because the TRAVIS_COMMIT is set to a merge commit sometimes, we can't
use it to checkout the mantl source clone for CI tests. We get the head
commit directly via shell commands, and add it to the env as
CI_HEAD_COMMIT. This is then used by subsequent commands.

I removed the `touch` commands with a far more useful call to ssh-add,
and I even kill the agent's process during the destroy phase.

Scrubbing:
- Moved non-OS code into `else` statements more consistently. This will
  speed up OS runs marginally
- Remove user-specific commands. This was cluttering up the code and my
  brain for little gain, since they aren't needed now.