Filter head commit by tested branch
We need to exclude other branches from the filter, because if there are
more recent merges in the repository, then there will be issues with the
wrong branch being pulled up. This error came up the first time I tried
this, because a different PR was merged into master just minutes after I
submitted this.