Add -s arg to kill, reduce nodes and destroy calls
I was trying to kill the ssh agent process without specifying what
signal to send. I've chosen SIGTERM for gentle destruction

I reduced the nunber of nodes that are created for OS clusters by 4
nodes.

Also, I've reduced the number of subproc calls for the destruction phase
by repeating the terraform command in the same call.