TRAVIS DOCKER Install bash for provisioning
There are some provisioning scripts used within ansible that require
bash syntax in order to run. Previously, I didn't install bash in the
container because we were trying to keep the image simpler. However,
if we don't include bash, that means we have to refactor scripts to be
sh compatible, and that is not a good trade-off.