[TRAVIS] Use remote tfstate to simplify CI ops
Our CI builds do not currently preserve state between builds. When
the destroy step doesn't trigger properly, infrastructure can hang
around, and that causes cascading quota failures in subsequent builds.
We could cache the tfstate within Travis, but that doesn't allow any
other tools access to the tfstate, which would be useful for external
tools. Therefore, we can use remote tfstate to allow for state to
persist, as well as expose the information to other tools we may create.

Supercedes and closes #1718