[TRAVIS] Replace fork check with proper logic
The code that I added in was the wrong code. It turns out that the logic
for the fork check was included in .travis.yml. When I found it, I also
realized that there was a much simpler solution, so I included that
instead. The code that I removed can be replaced with a safelist config
in .travis.yml