Update consul.cfg.j2
Local service is trying to use consul.service.consul versus 127.0.0.1, so changed to 0.0.0.0 so it gets picked up on the default external IP.  Given this file hasn't changed, I am not certain the entire repercussions of this.  As it stands though, deploying to AWS doesn't work without the change (again, there might be a better fix.)  Specifically, the UI is down, suspect due to lack of Consul access.